+++ stage3/application/src/main/java/com/neweducation/application/external/MockStudyDescriptionProviderService.java
@@ -7,7 +7,7 @@ package com.neweducation.application.external;
  */
 public class MockStudyDescriptionProviderService implements IStudyDescriptionProviderService {
     @Override
-    public String getStudyDescription(int id) {
+    public String getStudyDescription(long id) {
         return "http://www.wiz.pwr.wroc.pl/files/prv/id0/Informator/informator%20w-8_29.06.2015.pdf";
     }
 }
